Update index to have the correct dependencies

+++ fabfile.py
@@ -187,28 +187,29 @@ def install_project(platform='android',
             local('rm {0}/plugins/*'.format(asset_dir))
 
         # set up fieldtrip plugins
-        pobj = json.loads(open(os.sep.join((proj_home, 'plugins.json'))).read())
-        proot = os.sep.join((root, 'plugins'))
-        for plugin, details in pobj['plugins'].iteritems():
-            if len(details) == 0:
-                # plugin is from field trip
-                src = os.sep.join((proot, 'fieldtrip-plugins', plugin))
-            elif plugin[:3] == 'lib':
-                # TODO bower plugin
-                pass
-            else:
-                # git repository
-                src = os.sep.join((proot, plugin))
-                if not os.path.isdir(src):
-                    with lcd(proot):
-                        local('git clone {0} {1}'.format(details, plugin))
-
-            if os.path.isdir(src):
-                dest = os.sep.join((asset_dir, 'plugins', plugin))
-                local('ln -s {0} {1}'.format(src, dest))
-            else:
-                print 'No such plugin: {0}'.format(src)
-                exit(-1)
+        if os.path.exists(os.sep.join((proj_home, 'plugins.json'))):
+            pobj = json.loads(open(os.sep.join((proj_home, 'plugins.json'))).read())
+            proot = os.sep.join((root, 'plugins'))
+            for plugin, details in pobj['plugins'].iteritems():
+                if len(details) == 0:
+                    # plugin is from field trip
+                    src = os.sep.join((proot, 'fieldtrip-plugins', plugin))
+                elif plugin[:3] == 'lib':
+                    # TODO bower plugin
+                    pass
+                else:
+                    # git repository
+                    src = os.sep.join((proot, plugin))
+                    if not os.path.isdir(src):
+                        with lcd(proot):
+                            local('git clone {0} {1}'.format(details, plugin))
+    
+                if os.path.isdir(src):
+                    dest = os.sep.join((asset_dir, 'plugins', plugin))
+                    local('ln -s {0} {1}'.format(src, dest))
+                else:
+                    print 'No such plugin: {0}'.format(src)
+                    exit(-1)
 
         # set up bower dependecies
         for dep in bower['dependency_locations']:
